+++ inst/test-refs/read-output-refs/1_ext_ref_floor0.R
@@ -1,70 +1,72 @@
-structure(list(ITERATION = c(5, 10, 15, 20, 25, 28), THETA1 = c(
-  4.24454,
-  5.00791, 2.40082, 2.30528, 2.31302, 2.31034
-), THETA2 = c(
-  20.1268,
-  77.1104, 53.7872, 55.8065, 54.4961, 54.9596
-), THETA3 = c(
-  18.4447,
-  503.064, 382.758, 472.255, 465.37, 464.659
-), THETA4 = c(
-  -0.0060458,
-  -0.132147, -0.0769871, -0.0805144, -0.080792, -0.0805722
-), THETA5 = c(
-  5.26539,
-  6.25451, 4.03637, 4.12591, 4.13612, 4.1303
-), `SIGMA(1,1)` = c(
-  1,
-  1, 1, 1, 1, 1
-), `OMEGA(1,1)` = c(
-  0.0869929, 0.00708461, 0.116655,
-  0.0909862, 0.0961335, 0.0964407
-), `OMEGA(2,1)` = c(
-  0, 0, 0, 0,
-  0, 0
-), `OMEGA(2,2)` = c(
-  0.829972, 0.159955, 0.2438, 0.206974,
-  0.157375, 0.153571
-), OBJ = c(
-  3740.5412530896, 3133.06378054888,
-  2648.7566483156, 2638.50298364246, 2636.87787063403, 2636.84577032908
-)), class = c("spec_tbl_df", "tbl_df", "tbl", "data.frame"), row.names = c(
-  NA,
-  -6L
-), spec = structure(list(cols = list(ITERATION = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), THETA1 = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), THETA2 = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), THETA3 = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), THETA4 = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), THETA5 = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), `SIGMA(1,1)` = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), `OMEGA(1,1)` = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), `OMEGA(2,1)` = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), `OMEGA(2,2)` = structure(list(), class = c(
-  "collector_double",
-  "collector"
-)), OBJ = structure(list(), class = c(
-  "collector_double",
-  "collector"
-))), default = structure(list(), class = c(
-  "collector_guess",
-  "collector"
-)), delim = NULL, skip = 2L), class = "col_spec"))
+structure(list(
+  ITERATION = c(5, 10, 15, 20, 25, 30, 35, 37),
+  THETA1 = c(
+    6.43442, 5.37097, 2.07084, 2.29186, 2.36816, 2.32142,
+    2.31716, 2.31716
+  ), THETA2 = c(
+    5.76764, 61.4384, 60.2547,
+    56.1037, 55.6013, 54.3554, 54.6151, 54.6151
+  ), THETA3 = c(
+    14.4313,
+    255.291, 432.397, 340.312, 440.134, 464.034, 462.514, 462.514
+  ), THETA4 = c(
+    0.0240209, 0.120229, -0.0647548, -0.0774946,
+    -0.0801052, -0.0834912, -0.0820082, -0.0820082
+  ), THETA5 = c(
+    4.95595,
+    2.58218, 3.80554, 4.07723, 4.19512, 4.218, 4.17959, 4.17959
+  ), `SIGMA(1,1)` = c(1, 1, 1, 1, 1, 1, 1, 1), `OMEGA(1,1)` = c(
+    0.0624291,
+    0.045672, 0.307546, 0.144886, 0.0834166, 0.102911, 0.0985328,
+    0.0985328
+  ), `OMEGA(2,1)` = c(0, 0, 0, 0, 0, 0, 0, 0), `OMEGA(2,2)` = c(
+    0.964661,
+    0.220577, 0.18084, 0.219414, 0.198373, 0.158911, 0.156825,
+    0.156825
+  ), OBJ = c(
+    4186.09870907074, 3349.93513253719, 2610.86670752869,
+    2603.19925959281, 2585.99351850135, 2583.40361066657, 2583.31090094174,
+    2583.31090094174
+  )
+), class = c(
+  "spec_tbl_df", "tbl_df", "tbl",
+  "data.frame"
+), row.names = c(NA, -8L), spec = structure(list(
+  cols = list(ITERATION = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), THETA1 = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), THETA2 = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), THETA3 = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), THETA4 = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), THETA5 = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), `SIGMA(1,1)` = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), `OMEGA(1,1)` = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), `OMEGA(2,1)` = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), `OMEGA(2,2)` = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  )), OBJ = structure(list(), class = c(
+    "collector_double",
+    "collector"
+  ))), default = structure(list(), class = c(
+    "collector_guess",
+    "collector"
+  )), delim = NULL, skip = 2L
+), class = "col_spec"))